feat: file changes check all moved to platform agnostic
