+++ src/rules/node.test.ts
@@ -6,11 +6,8 @@ declare const global: any;
 
 describe('Node info', () => {
 
-  let warnMock: jest.Mock<any>;
-
   beforeEach(() => {
-    warnMock = jest.fn();
-    global.warn = jest.fn(warnMock);
+    global.warn = jest.fn();
     global.message = jest.fn();
     global.fail = jest.fn();
     global.markdown = jest.fn();
@@ -21,63 +18,6 @@ describe('Node info', () => {
     global.message = undefined;
     global.fail = undefined;
     global.markdown = undefined;
-  });
-
-  describe('Modified files', () => {
-
-    it('Should not warn when modified files are not on the to watch list', async () => {
-      global.danger = {
-        git: { modified_files: ['a random file', 'another random file'] },
-      };
-      await node.shouldNotHaveBeenChanged();
-
-      expect(global.warn).not.toBeCalled();
-
-    });
-
-    it('Should not warn when deleted files are not on the to watch list', async () => {
-      global.danger = {
-        git: { deleted_files: ['a random file', 'another random file'] },
-      };
-      await node.shouldNotHaveBeenChanged();
-
-      expect(global.warn).not.toBeCalled();
-
-    });
-
-    filesToCheck.forEach(file => {
-      it(`Checks if warns when ${file} is modified`, async () => {
-        global.danger = {
-          git: { modified_files: ['a random file', file, 'another random file'] },
-        };
-        await node.shouldNotHaveBeenChanged();
-
-        expect(global.warn).toBeCalled();
-
-        const expected = [
-          expect.stringMatching('The following files are rarely modified but were commited'),
-          expect.stringMatching(file),
-        ];
-        expect(warnMock.mock.calls[0]).toEqual(expect.arrayContaining(expected));
-      });
-
-      it(`Checks if warns when ${file} is deleted`, async () => {
-        global.danger = {
-          git: { deleted_files: ['a random file', file, 'another random file'] },
-        };
-        await node.shouldNotHaveBeenChanged();
-
-        expect(global.warn).toBeCalled();
-
-        const expected = [
-          expect.stringMatching('The following files are rarely modified but were commited'),
-          expect.stringMatching(file),
-        ];
-        expect(warnMock.mock.calls[0]).toEqual(expect.arrayContaining(expected));
-      });
-
-    });
-
   });
 
   describe('console.log', () => {
