+++ src/rules/node.ts
@@ -1,6 +1,5 @@
 // Provides dev-time type structures for  `danger` - doesn't affect runtime.
 import {DangerDSLType} from '../../node_modules/danger/distribution/dsl/DangerDSL';
-import { Scope } from '../rule.type';
 declare const danger: DangerDSLType;
 
 export declare function message(message: string): void;
@@ -8,9 +7,7 @@ export declare function warn(message: string): void;
 export declare function fail(message: string): void;
 export declare function markdown(message: string): void;
 
-// adapted from https://stackoverflow.com/a/37324915/429521
-const intersect = <T>(xs: T[], ys: T[]): T[] => xs.filter(x => ys.some(y => y === x));
-
+import { Scope } from '../rule.type';
 /**
  * Return true if there is any match with pattern on any file created or modified.
  * Return false otherwise.
@@ -34,39 +31,25 @@ const changedFilesContainsRegex = async (pattern: RegExp): Promise<boolean> => {
 
 export const filesToCheck = [
   'yarn.lock',
-  'npm-shrinkwrap.json',
+  'npm-debug.log',
+  'yarn-error.log',
   'docker-compose.yml',
-  'Procfile',
-  'node_modules',
-  'tasks/options/env.coffee',
   'tslint.json',
   'tsconfig.json',
   '.nvmrc',
+  'Procfile',
+  'npm-shrinkwrap.json',
   '.env',
   '.env.test',
   '.env.sample',
-  'nodemon.json',
-  'Dangerfile',
+  'env.coffee',
+  `nodemon.json`,
 ];
 
 /**
- * PR rules
+ * Node rules
  */
 export let node: Scope = {
-  /** Warns when ckey config files have been modified */
-  async shouldNotHaveBeenChanged() {
-
-    const files = ([] as string[]).concat(
-      danger.git.deleted_files  || [],
-      danger.git.modified_files || [],
-    );
-
-    const toWarn = intersect(filesToCheck, files);
-
-    if (toWarn.length > 0) {
-      warn(`The following files are rarely modified but were commited: ${toWarn.map(f => `'${f}'`).join(', ')}`);
-    }
-  },
 
   /** Warns when 'console.log' is added to code */
   async consoleLog() {
